Refactor maven repos and function locations
