+++ build.gradle.kts
@@ -5,6 +5,19 @@ plugins {
 allprojects {
     repositories {
         mavenCentral()
+        //add mvn.topobyte.de
+        maven {
+            url = uri("https://mvn.topobyte.de/")
+        }
+        maven {
+            url = uri("https://mvn.slimjars.com/")
+        }
+        maven {
+            url = uri("https://maven.pkg.jetbrains.space/public/p/ktor/eap")
+        }
+        maven {
+            url = uri("https://maven.pkg.jetbrains.space/kotlin/p/kotlin/dev")
+        }
     }
 }
